server/requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
colorama==0.4.6
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
googlemaps==4.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyparsing==3.2.1
PyPika==0.48.9
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.2
Werkzeug==3.1.3
